refactor: Rename NewCollectivePage to CollectivePage
