+++ components/collective-page/README.md
@@ -1,6 +1,6 @@
 # Collective page
 
-This file documents some the development choices made for the new collective page, as well
+This file documents some the development choices made for the collective page, as well
 as some conventions that we must follow.
 
 ### Sections
@@ -14,7 +14,7 @@ The binding of the data to the section happens in `components/collective-page/in
 The following rule should ideally be applied when fetching data for a section:
 
 If the section is going to be displayed for the majority of the collectives (contribute, updates...etc)
-then we should fetch the data in the page container, `pages/new-collective-page.js`.
+then we should fetch the data in the page container, `pages/collective-page.js`.
 
 However if the section will only be displayed for certain collectives, under certain conditions or
 if it will only be displayed for users/organizations/hosts then the data should be fetched at the component
