+++ test/cypress/integration/06-host.test.js
@@ -1,13 +1,13 @@
 describe('New host page', () => {
   /**
-   * About section is already tested in `04-collective_v2.test.js`
+   * About section is already tested in `04-collective.test.js`
    */
   before(() => {
-    cy.visit('/opensourceorg/v2');
+    cy.visit('/opensourceorg');
   });
 
   describe('Contributions section', () => {
-    // The rest of the contributions section is already tested in `05-user_v2.test.js`
+    // The rest of the contributions section is already tested in `05-user.test.js`
     it('Show fiscally hosted collectives', () => {
       cy.contains('[data-cy~="filter-button"]', 'Hosted Collectives').click();
       cy.contains('[data-cy=Contributions]', 'Open Source Collective');
@@ -25,7 +25,7 @@ describe('New host page', () => {
   });
 
   describe('Transactions section', () => {
-    // The rest of the transactions section tests are in `05-user_v2.test.js`
+    // The rest of the transactions section tests are in `05-user.test.js`
     it("Has no filters (because hosts don't create expenses)", () => {
       cy.contains('[data-cy~="filter-button"]', 'Expenses').should('not.exist');
     });
